[refactor] #11 OAuth 관련 기능 로직 분리
