+++ src/main/java/org/sopt/sweet/domain/member/entity/Member.java
@@ -9,7 +9,7 @@ import org.sopt.sweet.global.common.BaseTimeEntity;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
-@Builder(access = AccessLevel.PUBLIC)
+@Builder(access = AccessLevel.PUBLIC) //수정해야함
 @Table(name = "member")
 @Entity
 public class Member extends BaseTimeEntity {
@@ -30,4 +30,11 @@ public class Member extends BaseTimeEntity {
 
     private String nickName;
 
+    @Builder
+    public Member(Long socialId, String nickName, SocialType socialType, String profileImg) {
+        this.socialId = socialId;
+        this.nickName = nickName;
+        this.socialType = socialType;
+        this.profileImg = profileImg;
+    }
 }
